+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.0.38] - 08/24/2020
+
+* Update chat bubble to be able to replace the default
+reply widget, such as images etc.
+
 ## [0.0.37] - 08/21/2020
 
 * Update the extended_text_field package to 4.0.0